Update the "shell-variables" files
Include the "backgrounddim" key.  This is a colour value that is derived
programmatically as a median point between the two background values (bg
and bgalt).  It is not meant to be part of the standard 16-colour
palette, but only for use in special cases.

For technical details, read this entry in my website's codelog section:
https://protesilaos.com/codelog/2019-01-26-manipulating-colours/

The latest addition to the "shell-variables" template is in line with
the changes to the Tempus themes generator version 0.25.0.20190330:
https://gitlab.com/protesilaos/tempus-themes-generator